# pkg/loadbalancer allocation optimizations and cleanup