--- a/operator/watchers/service_sync.go
+++ b/operator/watchers/service_sync.go
@@ -221,7 +221,7 @@ func (d DefaultClusterServiceConverter) Convert(k8sService *slim_corev1.Service,
 		portConfig := serviceStore.PortConfiguration{}
 		for _, port := range k8sService.Spec.Ports {
 			p := loadbalancer.NewL4Addr(loadbalancer.L4Type(port.Protocol), uint16(port.Port))
-			portConfig[string(port.Name)] = p
+			portConfig[string(port.Name)] = &p
 		}
 		svc.Frontends = map[string]serviceStore.PortConfiguration{}
 		clusterIPs := k8sService.Spec.ClusterIPs