--- a/pkg/loadbalancer/redirectpolicy/redirectpolicy.go
+++ b/pkg/loadbalancer/redirectpolicy/redirectpolicy.go
@@ -225,7 +225,7 @@ func getSanitizedLocalRedirectPolicy(cfg Config, log *slog.Logger, name, namespa
 				return nil, fmt.Errorf("invalid address matcher port: %w", err)
 			}
 			// Set the scope to ScopeExternal as the externalTrafficPolicy is set to Cluster.
-			fe = *lb.NewL3n4Addr(proto, addrCluster, p, lb.ScopeExternal)
+			fe = lb.NewL3n4Addr(proto, addrCluster, p, lb.ScopeExternal)
 			feMappings[i] = feMapping{
 				feAddr: fe,
 				fePort: lb.FEPortName(pName),
@@ -260,7 +260,7 @@ func getSanitizedLocalRedirectPolicy(cfg Config, log *slog.Logger, name, namespa
 			}
 			// Set the scope to ScopeExternal as the externalTrafficPolicy is set to Cluster.
 			// frontend ip will later be populated with the clusterIP of the service.
-			fe = *lb.NewL3n4Addr(proto, cmtypes.AddrCluster{}, p, lb.ScopeExternal)
+			fe = lb.NewL3n4Addr(proto, cmtypes.AddrCluster{}, p, lb.ScopeExternal)
 			feMappings[i] = feMapping{
 				feAddr: fe,
 				fePort: lb.FEPortName(pName),