Last set of allocation optimizations for the new load-balancing control-plane:
* Store the `k8s.Backend.Ports` the other way around: `address -> port names` instead of `port name -> address` as that's the way we want to consume them. This avoids building temporary data structures in `pkg/loadbalancer/reflectors/k8s.go`
* Add `UpsertAndReleaseBackends` to reduce calls to `RefreshFrontends` and thus avoid some copies and insertions
* Change `NewL4Addr` and `NewL3n4Addr` to return a value instead of reference. These are so small that there's very little point in heap allocating them.
* Move `idAllocator` into `id_allocator.go` and stop using `StringID()` and instead index by `ID` and `L3n4Addr`.
* Specialize `idAllocator` for `ServiceID | BackendID`
* Remove unused types and code in pkg/loadbalancer using `deadcode`

`pkg/loadbalancer/benchmark` before:
```
Avg: Allocated 669602kB in total, 2708662 objects / 213388kB still reachable (per service:  54 objs, 13713B alloc,  4370B in-use)
Avg: Reconciled 50000 objects in 811.8553ms  (16.237µs  per service /  61588 services per second)
```

After:
```
Avg: Allocated 584072kB in total, 2379211 objects / 172144kB still reachable (per service:  47 objs, 11961B alloc,  3525B in-use)
Avg: Reconciled 50000 objects in 708.17076ms (14.163µs  per service /  70607 services per second)
```