@joamaki 2025-07-01
> I think as part of commit 6, the unused `lrpMetricsNoop` type and `NewLRPMetricsNoop` function could be removed as well:
> 
> ```diff
> diff --git a/pkg/loadbalancer/redirectpolicy/cell.go b/pkg/loadbalancer/redirectpolicy/cell.go
> index a090111abae0..b9351400ec87 100644
> --- a/pkg/loadbalancer/redirectpolicy/cell.go
> +++ b/pkg/loadbalancer/redirectpolicy/cell.go
> @@ -96,16 +96,3 @@ type LRPMetrics interface {
>         AddLRPConfig(loadbalancer.ServiceName)
>         DelLRPConfig(loadbalancer.ServiceName)
>  }
> -
> -type lrpMetricsNoop struct {
> -}
> -
> -func (p *lrpMetricsNoop) AddLRPConfig(loadbalancer.ServiceName) {
> -}
> -
> -func (p *lrpMetricsNoop) DelLRPConfig(loadbalancer.ServiceName) {
> -}
> -
> -func NewLRPMetricsNoop() LRPMetrics {
> -       return &lrpMetricsNoop{}
> -}
> ```

Yep makes sense. Removed.